Readme
This code is designed to check incoming data from Eltek equipment, in the form of .csv files. It is designed to check the incoming data every hour and email you if there is any missing data.

This code is designed to sit on a AWS EC2 and constantly check the data. The code will output to the console, telling you what it is doing and if there is any missing data. There is also a countdown timer, showing the minutes to the next 'run'.

The code will copy the original files from the Eltek GPRS folder and put them into a 'checking' sub folder. This is to avoid any access issues.

See the eltek_datachecker.py for dependancies. This works for sending emails from gmail - i have not tested other email providers.

You will need to edit:

1. the directory where the csv files are
2. the csv files you want to check
3. email recipient
4. email sender (gmail tested) and password
5. email message

If the code fails due to an IOerror, it will retry 10 times, before quiting and reporting an error.